add test without fixture
